Fix test_function.py
Previously, due to a typo in the second function definition,
`funcion1o` and `function1d` were both pointing to the same function
`function1`, therefore the entire TestCase was passing trivially
because it wasn't actually checking anithing meaningful.

It's much safer to explicitly pass the function to the decorator and
not use the decorator syntax, as this requires to duplicate the
function definition and is prone to error.

Here we are not testing that the decorator syntax is working: that is
the Python developers' responsibility.