Fix threading issues with register_post_import_hook
1. Add a missing lock to ImportHookFinder.find_spec(). It should have
the same lock as find_module() since it accesses _post_import_hooks.

2. Fix a deadlock between notify_module_loaded() and ImportHookFinder.
This occurs because the former is calling import hooks under
_post_import_hooks_lock. If a hook is making any imports itself while
another thread is also making imports, the thread can grab the import
lock and wait for _post_import_hooks_lock in ImportHookFinder, while the
hook already has _post_import_hooks_lock and is waiting for the import
lock. This results in a deadlock. The solution is to call the hooks
outside of _post_import_hooks_lock.

Same for register_post_import_hook() and  ImportHookFinder.

3. Also fix an issue if register_post_import_hook() was called after a
module was already imported but was later removed from sys.modules. This
also makes the logic simpler.